Many bugs are fixed.

+++ RestaurantApp/ViewModels/ArticleDetailsViewModel.cs
@@ -1,9 +1,13 @@
 ﻿using EntityFramework.Models;
+using Microsoft.EntityFrameworkCore;
 using Prism.Commands;
 using Prism.Mvvm;
 using Prism.Regions;
 using Prism.Services.Dialogs;
 using RestaurantApp.Services.Interface;
+using System.Collections.Generic;
+using System.Linq;
+using System.Windows;
 
 namespace RestaurantApp.ViewModels
 {
@@ -57,7 +61,7 @@ namespace RestaurantApp.ViewModels
             }
         }
 
-       public void ShowEditArticleDetail(ArticleDetails articleDetails)
+        public async void ShowEditArticleDetail(ArticleDetails articleDetails)
         {
             DialogParameters dialogParameters = new DialogParameters
             {
@@ -65,11 +69,25 @@ namespace RestaurantApp.ViewModels
             };
 
             _dialogService.ShowDialog("editArticleDetailDialog", dialogParameters, r => { });
+
+            Article = await _databaseService.GetArticleByID(articleDetails.ArticleID, new EFContext());
         }
 
         private async void DeleteArticleDetail(ArticleDetails articleDetails)
         {
             using EFContext efContext = new EFContext();
+
+            List<TableArticleQuantity> tableArticleQuantities = await _databaseService.GetTableArticleQuantityByArticleID(articleDetails.ArticleID, efContext);
+            List<TableArticleQuantity> tableArticleQuantitiesFiltered = tableArticleQuantities.Where(x => x is not SoldTableArticleQuantity).ToList();
+
+            if (tableArticleQuantitiesFiltered.Count != 0)
+            {
+                MessageBox.Show("You can't edit article details while it's used!", "Article details", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
+            efContext.ChangeTracker.Clear();
+
             await _databaseService.DeleteArticleDetails(articleDetails, efContext);
             RaisePropertyChanged(nameof(Article));
         }
